[Feature] Rename services to adapters
